% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerRelation.R
\name{encodeCoalition}
\alias{encodeCoalition}
\alias{decodeCoalition}
\title{Internal coalition representation}
\usage{
encodeCoalition(v, elements)

decodeCoalition(x, elements)
}
\arguments{
\item{v}{A vector of coalition members to be encoded into an integer value.}

\item{elements}{A vector of elements for the encoded bits to be mapped to.}

\item{x}{An encoded integer value to be transformed into a vector of elements.}
}
\description{
Transform a vector of coalition members into its internal integer representation used by the \code{SocialRanking} package.
}
\details{
Coalitions are internally stored as 32-bit integers, where each bit signalizes if a player (or element) partakes in the given coalition or not.
For instance, a \verb{14 = (1110)_2} represents a coalition containing the elements \code{2}, \code{3}, and \code{4}.

By default, it is assumed that elements are a sequence of numbers starting from \code{1}.
If this is not the case, pass a supplementary \code{elements} parameter to map elements to their corresponding sequence indexes.

Since all coalitions passed to a \code{SocialRanking} functions are converted to integers, this processing step can be circumvented if the \code{asBit} flag in those functions are set to \code{TRUE} and the corresponding integer representation of the coalition is supplied.
}
\examples{
stopifnot(11 == encodeCoalition(c(1, 2, 4)))
stopifnot(11 == encodeCoalition(c('a', 'b', 'd'), elements = letters))

stopifnot(identical(decodeCoalition(11), c(1, 2, 4)))
stopifnot(identical(decodeCoalition(11, elements = letters), c('a', 'b', 'd')))

# potential usage
PowerRelation(list(list(7, 3, 1), list(6, 4), list(5, 2, 1, 0)), asBits = TRUE)

PowerRelation(list(list(7, 3, 1), list(6, 4), list(5, 2, 1, 0)), elements = letters[1:3], asBits = TRUE)
}
