% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerRelationEqs.R
\name{eqClasses}
\alias{eqClasses}
\alias{print.eqClasses}
\alias{[.eqClasses}
\alias{[[.eqClasses}
\alias{length.eqClasses}
\alias{as.list.eqClasses}
\alias{is.eqClasses}
\alias{sort.eqClasses}
\title{Internal class for equivalence classes of coalitions}
\usage{
\method{print}{eqClasses}(x, ...)

\method{[}{eqClasses}(x, i, ...)

\method{[[}{eqClasses}(x, i, ...)

\method{length}{eqClasses}(x)

\method{as.list}{eqClasses}(x, ...)

is.eqClasses(x, ...)

\method{sort}{eqClasses}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{An \code{eqClasses} object.}

\item{...}{Additional arguments (e.g., passed through to indexing methods).}

\item{i}{Index or indices.}

\item{decreasing}{Logical, used in \code{sort}.}

\item{asBits}{Logical flag. If \code{TRUE}, returns raw integer bitmasks. Otherwise, returns player vectors.}
}
\value{
Varies by method: can return an integer, list, or \code{eqClasses} object.
}
\description{
This object represents a collection of equivalence classes used internally to group coalitions.
Each class is a list of coalitions, encoded either as integers (bitmasks) or player vectors.
}
\details{
Users typically do not need to interact with this class directly, but several methods are defined for internal or advanced use.
}
\keyword{internal}
