% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRelation.R
\name{equivalenceClassIndexes}
\alias{equivalenceClassIndexes}
\title{Get all equivalence class indexes}
\usage{
equivalenceClassIndexes(powerRelation, coalition)
}
\arguments{
\item{powerRelation}{A \code{PowerRelation} object created by \code{\link[=newPowerRelation]{newPowerRelation()}}}

\item{coalition}{a coalition vector or \code{\link[sets:set]{sets::set}} that is part of \code{powerRelation}}
}
\value{
Numeric vector contatining indexes of \code{powerRelation$equivalenceClasses} containing \code{coalitions}.
Empty if none exist.
}
\description{
For one or more \code{coalitions}, return all equivalence class indexes they appear in.
}
\examples{
pr <- newPowerRelation(c(1,2), ">", c(1), "~", c(2))

}
\seealso{
Other equivalence class lookup functions: 
\code{\link{coalitionsAreIndifferent}()},
\code{\link{equivalenceClassIndex}()}
}
\concept{equivalence class lookup functions}
