% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePowerRelationTotal.R
\name{makePowerRelationTotal}
\alias{makePowerRelationTotal}
\title{Make Power Relation total}
\usage{
makePowerRelationTotal(powerRelation, includeEmptySet = TRUE)
}
\arguments{
\item{powerRelation}{A \code{\link{PowerRelation}} object created by \code{\link[=PowerRelation]{PowerRelation()}} or \code{\link[=as.PowerRelation]{as.PowerRelation()}}}

\item{includeEmptySet}{If \code{TRUE}, include the empty set in the last equivalence class if it is missing from the power relation.}
}
\value{
\code{\link{PowerRelation}} object containing the following values:
\itemize{
\item \verb{$elements}: vector of elements
\item \verb{$eqs}: equivalence classes. Nested list of lists, each containing vectors representing groups of elements in the same equivalence class
\item \verb{$coalitionLookup}: \verb{function(v)} taking a coalition vector \code{v} and returning the equivalence class it belongs to. See \code{\link[=coalitionLookup]{coalitionLookup()}} for more.
\item \verb{$elementLookup}: \verb{function(e)} taking an element \code{e} and returning a list of 2-sized tuples. See \code{\link[=elementLookup]{elementLookup()}} for more.
}
}
\description{
Append an equivalence to a power relation with all its missing coalitions to make it total.
}
\details{
A power relation is total if

\deqn{S \succeq T\text{ or }T \succeq S}{S>=T or T>=S}

for every \eqn{S, T \subseteq N}{S, T subset or equal to N}.

In other words, we can compare every coalition against every other coalition there is.
Given the equivalence classes of a \code{\link{PowerRelation}} object,
the function attaches a new equivalence class containing all the missing coalitions that make the given power relation total.
}
\examples{
pr <- as.PowerRelation(list(c(1,2), 3))
# 12 > 3

makePowerRelationTotal(pr)
# 12 > 3 > (123 ~ 13 ~ 23 ~ 1 ~ 2 ~ {})

makePowerRelationTotal(pr, includeEmptySet = FALSE)
# 12 > 3 > (123 ~ 13 ~ 23 ~ 1 ~ 2)

}
\seealso{
Other helper functions transorming existing \code{\link{PowerRelation}} objects: 
\code{\link{makePowerRelationMonotonic}()}
}
\concept{helper functions transorming existing \code{\link{PowerRelation}} objects}
