% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRelation.R
\name{PowerRelation.default}
\alias{PowerRelation.default}
\title{PowerRelation object}
\usage{
\method{PowerRelation}{default}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to or from other methods}
}
\value{
No return value.
}
\description{
Use \code{\link[=newPowerRelation]{newPowerRelation()}} or \code{\link[=newPowerRelationFromString]{newPowerRelationFromString()}} to create a PowerRelation object.
}
