% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socialranking-package.R
\docType{package}
\name{socialranking-package}
\alias{socialranking}
\alias{socialranking-package}
\title{socialranking: A package for constructing ordinal power relations and evaluating social ranking solutions}
\description{
The package \code{socialranking} offers functions to represent ordinal
information of coalitions and calculate the power relation between elements.
}
\details{
Use \code{browseVignettes("socialranking")} for more information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jassler/socialranking}
  \item Report bugs at \url{https://github.com/jassler/socialranking/issues}
}

}
\author{
\strong{Maintainer}: Felix Fritz \email{felix.fritz@dauphine.eu}

Authors:
\itemize{
  \item Jochen Staudacher \email{jochen.staudacher@hs-kempten.de} [copyright holder, thesis advisor]
  \item Moretti Stefano \email{stefano.moretti@lamsade.dauphine.fr} [copyright holder, thesis advisor]
}

}
\keyword{internal}
