% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{package}
\name{socialranking}
\alias{socialranking}
\title{socialranking: A package for constructing ordinal power relations and evaluating social ranking solutions}
\description{
The package \code{socialranking} offers functions to represent ordinal
information of coalitions and calculate the power relation between elements or players.
}
\details{
\code{\link[=newPowerRelation]{newPowerRelation()}} creates a \code{PowerRelation} object. \code{\link[=createPowerset]{createPowerset()}}
is a convenient function to generate a \code{\link[=newPowerRelation]{newPowerRelation()}} function call
for all possible coalitions.

For every function that calculates the score of a given element in
a power relation, there is a \code{ranking}-equivalent function that
displays the ranking between elements. These include:\tabular{ll}{
   \emph{Score Function} \tab \emph{Ranking Function} \cr
   \code{\link[=ordinalBanzhafScores]{ordinalBanzhafScores()}} \tab \code{\link[=ordinalBanzhafRanking]{ordinalBanzhafRanking()}} \cr
   \code{\link[=copelandScores]{copelandScores()}} \tab \code{\link[=copelandRanking]{copelandRanking()}} \cr
   \code{\link[=kramerSimpsonScores]{kramerSimpsonScores()}} \tab \code{\link[=kramerSimpsonRanking]{kramerSimpsonRanking()}} \cr
   \code{\link[=lexcelScores]{lexcelScores()}} \tab \code{\link[=lexcelRanking]{lexcelRanking()}} \cr
    \tab \code{\link[=dualLexcelRanking]{dualLexcelRanking()}} \cr
   \code{\link[=cumulativeScores]{cumulativeScores()}} \tab \code{\link[=cumulativelyDominates]{cumulativelyDominates()}}* \cr
    \tab \code{\link[=dominates]{dominates()}}* \cr
    \tab \code{\link[=cpMajorityComparison]{cpMajorityComparison()}}* \cr
}


*These functions only compare two elements and doen't rank them.

\code{\link[=powerRelationMatrix]{powerRelationMatrix()}} uses \code{\link[relations:relation]{relations::relation()}} to create
an incidence matrix between all competing coalitions. The incidence
matrix can be displayed with \code{\link[relations:incidence]{relations::relation_incidence()}}.
}
