% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingBoilerplate.R
\name{SocialRankingSolution.default}
\alias{SocialRankingSolution.default}
\title{\code{SocialRankingSolution} object}
\usage{
\method{SocialRankingSolution}{default}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to or from other methods}
}
\value{
No return value.
}
\description{
Use \code{\link[=doRanking]{doRanking()}} to create a \code{SocialRankingSolution} object.
}
