% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asPowerRelation.R
\name{as.PowerRelation}
\alias{as.PowerRelation}
\alias{as.PowerRelation.character}
\alias{as.PowerRelation.list}
\title{Create PowerRelation object}
\usage{
as.PowerRelation(x, ...)

\method{as.PowerRelation}{character}(x, ...)

\method{as.PowerRelation}{list}(x, ..., comparators = c(">"))
}
\arguments{
\item{x}{An object}

\item{...}{Optional additional parameters}

\item{comparators}{Vector of ">" or "~" characters}
}
\description{
Alternative ways of creating \code{\link{PowerRelation}} objects.
}
\section{Using a character string}{


The same way a power relation \eqn{\succeq}{>= }
}

\section{Using a list}{


Man man man.
}

