% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRelation.R
\name{coalitionsAreIndifferent}
\alias{coalitionsAreIndifferent}
\title{Are coalitions indifferent}
\usage{
coalitionsAreIndifferent(powerRelation, c1, c2)
}
\arguments{
\item{powerRelation}{A \code{\link{PowerRelation}} object created by \code{\link[=newPowerRelation]{newPowerRelation()}}}

\item{c1}{Coalition \link[base:c]{vector} or \code{\link[sets:set]{sets::set()}}}

\item{c2}{Coalition \link[base:c]{vector} or \code{\link[sets:set]{sets::set()}}}
}
\value{
Logical value \code{TRUE} if \code{c1} and \code{c2} are in the same equivalence class, else \code{FALSE}.
}
\description{
Check if coalitions are indifferent from one another, or, if they appear in the same
equivalence class.
}
\examples{
pr <- newPowerRelation(c(1,2), ">", c(1), "~", c(2))

stopIfNot(F == coalitionsAreIndifferent(pr, c(1,2), c(1)))
stopIfNot(T == coalitionsAreIndifferent(pr, 2, 1))

# Note that it doesn't fail with non-existing power relations
stopIfNot(F == coalitionsAreIndifferent(pr, 1, c()))
stopIfNot(T == coalitionsAreIndifferent(pr, 3, c(1,2,3)))

}
\seealso{
Other equivalence class lookup functions: 
\code{\link{equivalenceClassIndex}()}
}
\concept{equivalence class lookup functions}
