% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPowerset.R
\name{createPowerset}
\alias{createPowerset}
\title{Create powerset}
\usage{
createPowerset(
  elements,
  includeEmptySet = TRUE,
  result = c("return", "print", "printCompact", "copy", "copyCompact")
)
}
\arguments{
\item{elements}{vector of elements}

\item{includeEmptySet}{If \code{TRUE}, an empty vector is added at the end}

\item{result}{What to do with the result. Can be either:
\itemize{
\item \code{"return"}: return list object
\item \code{"print"}: create valid string to call \code{\link[=PowerRelation]{PowerRelation()}} or \code{\link[=as.PowerRelation]{as.PowerRelation()}} and print it
\item \code{"copy"}: create valid string to call \code{\link[=PowerRelation]{PowerRelation()}} or \code{\link[=as.PowerRelation]{as.PowerRelation()}} and copy it to clipboard
\item \code{"printCompact"} and \code{"copyCompact"}: same as \code{"print"} and \code{"copy"} but without newlines
}}
}
\value{
List of power set vectors.
If the parameter \code{result} is set to \code{"print"} or \code{"copy"}, nothing is returned.
Instead, a character string is generated that can be used in R to call and create a new \code{\link{PowerRelation}} object.
This string is either printed or copied to clipboard (see argument \code{result}).
}
\description{
Given a vector of elements generate a power set.
}
