% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRelation.R
\name{elementLookup}
\alias{elementLookup}
\title{Element lookup}
\usage{
elementLookup(powerRelation, element)
}
\arguments{
\item{powerRelation}{A \code{\link{PowerRelation}} object created by \code{\link[=PowerRelation]{PowerRelation()}} or \code{\link[=as.PowerRelation]{as.PowerRelation()}}}

\item{element}{an element in \code{powerRelation$elements}}
}
\value{
List of 2-value tuples.
First value of a tuple indicates the equivalence class index, the second value the index inside that equivalence class with the coalition containing the element.
See \code{elementLookup()} in \code{\link[=PowerRelation]{PowerRelation()}} for more information.
}
\description{
List coalitions that an element appears in.
}
\details{
This function basically just calls \code{powerRelation$elementLookup(element)}.
}
\examples{
pr <- as.PowerRelation("12 > 2 ~ 1")

l <- elementLookup(pr, 1)
l
# (1,1), (2,2)

l |> sapply(function(tuple) 1 \%in\% pr$eqs[[tuple[1]]][[tuple[2]]]) |> all() |> stopifnot()

# if element does not exist, it returns NULL
pr |> elementLookup(3) |> is.null() |> stopifnot()

}
\seealso{
Other lookup functions: 
\code{\link{coalitionsAreIndifferent}()},
\code{\link{equivalenceClassIndex}()}
}
\concept{lookup functions}
