% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerRelation.R
\name{equivalenceClassIndex}
\alias{equivalenceClassIndex}
\title{Get index of equivalence class containing a coalition}
\usage{
equivalenceClassIndex(...)
}
\arguments{
\item{...}{Any parameter}
}
\value{
No return value.
}
\description{
Deprecated. Use powerRelation$coalitionLookup() instead.
}
