% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRelation.R
\name{equivalenceClassIndex}
\alias{equivalenceClassIndex}
\alias{coalitionLookup}
\title{Get index of equivalence class containing a coalition}
\usage{
equivalenceClassIndex(powerRelation, coalition)

coalitionLookup(powerRelation, coalition)
}
\arguments{
\item{powerRelation}{A \code{\link{PowerRelation}} object created by \code{\link[=PowerRelation]{PowerRelation()}} or \code{\link[=as.PowerRelation]{as.PowerRelation()}}}

\item{coalition}{a coalition \link[base:c]{vector} or that is part of \code{powerRelation}}
}
\value{
Numeric value, equivalence class index where \code{coalition} appears in.
\code{-1} if the coalition does not exist.
See \code{coalitionLookup()} in \code{\link[=PowerRelation]{PowerRelation()}} for more information.
}
\description{
Given a \code{coalition} \link[base:c]{vector}, return the equivalence class index it appears in.
}
\details{
This function basically just calls \code{powerRelation$coalitionLookup(coalition)}.

\code{equivalenceClassIndex()} serves as an alias to \code{coalitionLookup()}.
}
\examples{
pr <- as.PowerRelation("12 > 2 ~ 1")

equivalenceClassIndex(pr, c(1, 2))
# 1

equivalenceClassIndex(pr, c(1))
# 2

equivalenceClassIndex(pr, c(2))
# 2

equivalenceClassIndex(pr, c())
# -1 <- empty set does not exist

}
\seealso{
Other lookup functions: 
\code{\link{coalitionsAreIndifferent}()},
\code{\link{elementLookup}()}
}
\concept{lookup functions}
