% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexcel1.R
\name{L1Scores}
\alias{L1Scores}
\title{L1 Scores}
\usage{
L1Scores(powerRelation, elements = NULL)
}
\arguments{
\item{powerRelation}{A \code{\link{PowerRelation}} object created by \code{\link[=PowerRelation]{PowerRelation()}} or \code{\link[=as.PowerRelation]{as.PowerRelation()}}}

\item{elements}{vector of elements of which to calculate their scores.
If \code{elements == NULL}, create vectors for all elements in \code{pr$elements}}
}
\value{
Score function returns a list of type \code{Lexcel1Scores} and length of \code{powerRelation$elements}
(unless parameter \code{elements} is specified).
Each index contains a vector of length \code{powerRelation$eqs}, the number of
times the given element appears in each equivalence class.
}
\description{
Calculate the L1 (or Lexcel1) scores.
}
\details{
Similar
}
\references{
\insertRef{2019Lexcel}{socialranking}

\insertRef{2021Lexcel}{socialranking}
}
\seealso{
Other score vector functions: 
\code{\link{copelandScores}()},
\code{\link{cumulativeScores}()},
\code{\link{kramerSimpsonScores}()},
\code{\link{lexcelScores}()},
\code{\link{ordinalBanzhafScores}()}
}
\concept{score vector functions}
