% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerRelation.R
\name{newPowerRelation}
\alias{newPowerRelation}
\title{New Power Relation}
\usage{
newPowerRelation(...)
}
\arguments{
\item{...}{Any parameter.}
}
\value{
No return value.
}
\description{
Deprecated. Use \code{\link[=PowerRelation]{PowerRelation()}} instead.
}
