% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerRelation.R
\name{newPowerRelationFromString}
\alias{newPowerRelationFromString}
\title{New \code{\link{PowerRelation}} object}
\usage{
newPowerRelationFromString(...)
}
\arguments{
\item{...}{Any parameter.}
}
\value{
No return value.
}
\description{
Deprecated. Use \code{\link[=as.PowerRelation]{as.PowerRelation()}} instead.
}
