% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerRelation.R
\name{PowerRelation}
\alias{PowerRelation}
\alias{is.PowerRelation}
\alias{print.PowerRelation}
\title{PowerRelation object}
\usage{
PowerRelation(x, ...)

is.PowerRelation(x, ...)

\method{print}{PowerRelation}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Arguments passed to or from other methods}
}
\value{
No return value.
}
\description{
Use \code{\link[=newPowerRelation]{newPowerRelation()}} or \code{\link[=newPowerRelationFromString]{newPowerRelationFromString()}} to create a PowerRelation object.
}
