% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{powerRelationGenerator}
\alias{powerRelationGenerator}
\title{Generate power relations}
\usage{
powerRelationGenerator(coalitions)
}
\arguments{
\item{coalitions}{List of coalition vectors. An empty coalition can be set with \code{c()}.}
}
\value{
A generator function.
Every time this generator function is called, a different \code{\link{PowerRelation}} object is returned.
Once all possible power relations have been generated, the generator function returns \code{NULL}.
}
\description{
Based on a list of coalitions, create a generator function that returns a new \code{\link{PowerRelation}} object with every call.
\code{NULL} is returned once every possible power relation has been generated.
}
\examples{
coalitions <- createPowerset(c('a','b'), includeEmptySet = FALSE)
# list(c('a','b'), 'a', 'b')

gen <- powerRelationGenerator(coalitions)

pr <- gen()
while(!is.null) {
  print(pr)
  pr <- gen()
}
# (ab ~ a ~ b)
# (ab ~ a) > b
# (ab ~ b) > a
# (a ~ b) > ab
# ab > (a ~ b)
# a > (ab ~ b)
# b > (ab ~ a)
# ab > a > b
# ab > b > a
# a > ab > b
# b > ab > a
# a > b > ab
# b > a > ab
# NULL

}
