% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{testRelation}
\alias{testRelation}
\alias{\%:\%}
\alias{\%>dom\%}
\alias{\%>cumuldom\%}
\alias{\%>cp\%}
\alias{\%>banz\%}
\alias{\%>cop\%}
\alias{\%>ks\%}
\alias{\%>lex\%}
\alias{\%>duallex\%}
\title{Test relation between two elements}
\usage{
testRelation(powerRelation, e1)

powerRelation \%:\% e1

pr_e1 \%>dom\% e2

pr_e1 \%>cumuldom\% e2

pr_e1 \%>cp\% e2

pr_e1 \%>banz\% e2

pr_e1 \%>cop\% e2

pr_e1 \%>ks\% e2

pr_e1 \%>lex\% e2

pr_e1 \%>duallex\% e2
}
\arguments{
\item{powerRelation}{A \code{PowerRelation} object created by \code{\link[=newPowerRelation]{newPowerRelation()}}}

\item{e1, e2}{Elements in \code{powerRelation$elements}}

\item{pr_e1}{\code{PowerRelation} and \code{e1} element, packed into a list using \code{pr \%:\% e1}}
}
\description{
On a given \code{PowerRelation} object \code{pr}, check if \code{e1} relates to \code{e2} in a given social ranking solution.
}
\details{
The function \code{testRelation} is somewhat only used to make the offered comparison operators in the package better discoverable.

\code{testRelation(pr, e1)} is equivalent to \code{pr \%:\% e1} and \code{list(pr, e1)}. It should be used together with one of the
comparison operators listed in the usage section.
}
\examples{
pr <- newPowerRelationFromString(
  "123 > 12 ~ 13 > 3 > 1 ~ 2", asWhat = as.numeric
)

# Dominance
stopifnot(pr \%:\% 1 \%>dom\% 2)

# Cumulative dominance
stopifnot(pr \%:\% 1 \%>cumuldom\% 2)

# CP-Majority relation
stopifnot(pr \%:\% 1 \%>cp\% 2)

# Ordinal banzhaf relation
stopifnot(pr \%:\% 1 \%>banz\% 2)

# Copeland-like method
stopifnot(pr \%:\% 1 \%>cop\% 2)
stopifnot(pr \%:\% 2 \%>cop\% 1)

# Kramer-Simpson-like method
stopifnot(pr \%:\% 2 \%>ks\% 1)

# Lexicographical and Dual Lexicographical Excellence
stopifnot(pr \%:\% 1 \%>lex\% 3)
stopifnot(pr \%:\% 3 \%>duallex\% 1)

}
